requirements.txt
pandas~=2.2.1
aiogram~=3.4.1
beautifulsoup4~=4.12.3
selenium~=4.18.1
python-dotenv~=1.0.1